Tim noted a circular dependence w.r.t. bug #161.  Let's say the formula is
  x = 10  &&  x < 1

as it is in the minimized uart SMT example attached to the bug report.

The first atom, x = 10, is converted to a SAT literal (say "l").  The
(unit) clause "l" is asserted, causing MiniSat to do a round of
propagations, which does an arith theory check (fine), and an arith
theory propagate.  The arithmetic theory propagates x < 1, naturally,
and that caused an assertion failure since x < 1 doesn't yet have an
assigned SAT literal.

Detailed changelog:

* Moved TseitinCnfStream::handleAtom() to CnfStream::convertAtom().
* Added a second (boolean) argument "create" to
  CnfStream::getLiteral().  When "create" is true, getLiteral() will
  create the mapping for the atom passed in.  When "create" is false
  (the default), the assertion will fail (in debug builds)
* changed theory propagation (SatSolver::theoryPropagate()) to call
  getLiteral() with create == true.

Closes bug 161 on this branch.